Move util `Env` class closer to its single usage
`Env` class is partially copied from CDT and has very limited usage.
Therefore, it should not be a part of API and it is better to move it to
internal package, closer to its single usage.
